change dropout input type

+++ train.py
@@ -367,7 +367,7 @@ if __name__ == "__main__":
     parser.add_argument('--number_of_characters', type=int, default=69)
     parser.add_argument('--extra_characters', type=str, default='')
     parser.add_argument('--max_length', type=int, default=150)
-    parser.add_argument('--dropout_input', type=int, default=0.1)
+    parser.add_argument('--dropout_input', type=float, default=0.1)
     parser.add_argument('--epochs', type=int, default=10)
     parser.add_argument('--batch_size', type=int, default=128)
     parser.add_argument('--optimizer', type=str,
